MetaImage can be set as Visual element in topic article
